+++ src/main/java/com/github/alexthe666/iceandfire/entity/ai/AmphithereAIFollowOwner.java
@@ -91,7 +91,7 @@ public class AmphithereAIFollowOwner extends EntityAIBase {
                         for (int l = 0; l <= 4; ++l) {
                             for (int i1 = 0; i1 <= 4; ++i1) {
                                 if ((l < 1 || i1 < 1 || l > 3 || i1 > 3) && this.isTeleportFriendlyBlock(i, j, k, l, i1)) {
-                                    this.ampithere.setLocationAndAngles((double) ((float) (i + l) + 0.5F), (double) k, (double) ((float) (j + i1) + 0.5F), this.ampithere.rotationYaw, this.ampithere.rotationPitch);
+                                    this.ampithere.setLocationAndAngles((float) (i + l) + 0.5F, k, (float) (j + i1) + 0.5F, this.ampithere.rotationYaw, this.ampithere.rotationPitch);
                                     ampithere.getNavigator().clearPath();
                                     return;
                                 }
