+++ src/main/java/com/github/alexthe666/iceandfire/entity/ai/AquaticAIGetOutOfWater.java
@@ -58,7 +58,7 @@ public class AquaticAIGetOutOfWater extends EntityAIBase {
         for (int i = 0; i < 10; ++i) {
             BlockPos blockpos1 = blockpos.add(random.nextInt(20) - 10, random.nextInt(6) - 3, random.nextInt(20) - 10);
             if (this.world.getBlockState(blockpos1).isOpaqueCube()) {
-                return new Vec3d((double) blockpos1.getX(), (double) blockpos1.getY(), (double) blockpos1.getZ());
+                return new Vec3d(blockpos1.getX(), blockpos1.getY(), blockpos1.getZ());
             }
         }
